Change KES signature representation and serialization format.
This is so that we can reduce the size of KES signatures: instead of
duplicating *both* VerKeys in each branch node, we only duplicate the
one we need, and reconstruct the other from the level below.

For this implementation, we now split the verification process into two steps:

1. verifyDSIGN on the 'sigma' part of the signature. We include the
   VerKey for the leaf node in the signature itself, and just verify
   against that. This is not sufficient as validation on its own,
   hence...
2. ...recursively calculate the root VerKey along the Merkle path down
   to the leaf node, and compare it against the VerKey to verify
   against.

The advantage of this is that we only ever calculate the hashes we
absolutely need, and avoid duplicate calculation efforts.